Split requests and humanity into separate modules.
